VE should set provisioned flag
There is a case in which VE may skip validation on assignments
in which all hosts are validated, but the provisioned flag was
never set to True.  This can happen if the move and rebuild of a
host has an issue, in which case provisioned at the assignment
level is never set as expected.

Fixes: https://github.com/redhat-performance/quads/issues/361
Change-Id: Id5099e2291ced36a5f46b47d52f619ece9cddac0