Fix timestamps for move-and-rebuild log
The example cron puts the date command inside of flock causing
the output of the command to be executed prior to running the
cron entry.  Changed the command examples to add Start and Finish
into each log file, as well as allign the move-and-rebuild example
to match the others.  This also fixes the issue where the finished
timestamp for move-and-rebuild to be incorrect.

Fixes: https://github.com/redhat-performance/quads/issues/555
Change-Id: I4bca7af90c5b180a54afb1e29fbe666f383da0f4