Set port speed under interfaces
This patch covers a fix on juniper connected interfaces
to explicitly set the juniper connected port speed to the
speed defined in the quads database for the interface(s)
of a host.

If the interface speed is set to 0, 1, or -1, then there
will be no additional setting.  Otherwise, speed is looked
up and divided by 1000.  For example, if the speed on the
quads interface is "25000", then the speed set on the interface
will add:

   set interfaces <interface name> speed 25g

Fixes: https://github.com/redhat-performance/quads/issues/566
Change-Id: I7c2d3788bb495eea6d08dbd43c4451e01c570823