Send emails in HTML format
This change allows for messages sent via the notify_tenant.py
tool to be sent via HTML by treating the message content in the
template passed to the tool as markdown.  This allows for any
included tables in the templates to be properly converted and
rendered in any email client that supports HTML.

For example, templates can include tables such as:

| *Header 1*  | *Header 2*  | *Header 3*  |
|-------------|-------------|-------------|
| row 1 col 1 | row 1 col 2 | row 1 col 3 |
| row 2 col 1 | row 2 col 2 | row 2 col 3 |
| row 3 col 1 | row 3 col 2 | row 3 col 3 |
| row 4 col 1 | row 4 col 2 | row 4 col 3 |

and emails will be properly formatted to show the tables.
Furthermore, JIRA has some basic support and intelligence
to render tables.  However, the lines that begin with "|-"
need to be stripped out, otherwise the posted messages will
have extra cruft and the tables will not render properly.

Fixes: https://github.com/redhat-performance/quads/issues/627
Change-Id: I10a16b2e6b0645f481699b3fd82b45b82dddafc1