Adjust templates
To make use of the updated postman python
(which uses markdown to html), make sure that
templates force newlines explicitly.  This is
done in "markdown" by ensuring lines for which
you wish to have newline end with 2 or more
spaces followed by a newline.  In other words,
if you want a message that include:

This is line 1.
This is line 2.

Then the template should have:

This is line 1.<SPACE><SPACE><NEWLINE>
This is line 2.

Two newlines already correctly starts a new
paragraph in the conversion of markdown to
html.

Change-Id: I79d62d4b2ca98aef8408cc624d63db273631d349